add new internalInputValueToAst method to Coercing to convert an internal input value back to an AST

+++ src/test/groovy/graphql/TestUtil.groovy
@@ -259,7 +259,7 @@ class TestUtil {
             def builder = newDirective().name(directiveName)
 
             names.each { argName ->
-                builder.argument(newArgument().name(argName).type(GraphQLInt).value(BigInteger.valueOf(0)).build())
+                builder.argument(newArgument().name(argName).type(GraphQLInt).value(0).build())
             }
             return builder.build()
         }.toArray() as GraphQLDirective[]
